requirements-dev.txt
--index-url https://pypi.devinfra.sentry.io/simple

docker>=3.7.0,<3.8.0
exam>=0.10.6
freezegun>=1.1.0
honcho>=1.1.0
openapi-core>=0.14.2
pytest>=7.2
pytest-cov>=4.0.0
pytest-django>=4.4.0
pytest-rerunfailures>=10
pytest-sentry>=0.1.10
pytest-xdist>=3
responses>=0.21.0

# pre-commit dependencies
pre-commit>=2.18.1
black>=22.3.0
flake8>=5.0.2
flake8-bugbear>=22.7.1
pyupgrade>=3
isort>=5.10.1

# For tools/. To be moved into redistributable dev environments.
# note: wheel is pulled in by pip-tools
pip-tools>=6.7.0
packaging>=21.3

# for type checking
django-stubs
lxml-stubs
msgpack-types>=0.2.0
mypy>=0.982
types-beautifulsoup4
types-freezegun
types-jsonschema
types-psycopg2
types-python-dateutil
types-pytz
types-pyyaml
types-redis
types-requests
types-simplejson>=3.17.7.2
